rpi_zero/requirements.txt
Flask
bokeh
requests
flask_sqlalchemy
pyserial
sqlaclhemy

